+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added "Certifications and Certified Courses" section
+- Rephrasing on multiple entries
 - Added secret switch to enable/disable the display of the full CV URL
 - Added version number to the CV page
 - Modified download button dimensions